+++ src/views/ApplicationPage/utils/captureApplication.ts
@@ -1,19 +1,34 @@
 import html2canvas from 'html2canvas';
 
-export const captureApplication = async (): Promise<File> => {
-  let imgUrl = '';
-  html2canvas(document.getElementById('applcationImg')!).then((canvas) => {
-    imgUrl = canvas.toDataURL();
-  });
-
+export const captureApplication = async () => {
   try {
-    const response = await fetch(imgUrl);
-    const blob = await response.blob();
-    const file = new File([blob], 'captureApplication.png', { type: 'image/png' });
+    const canvas = await html2canvas(document.getElementById('applicationImg')!);
+    if (canvas) {
+      console.log(canvas);
+      const imgUrl = canvas.toDataURL();
+      downloadURL(imgUrl, 'test.png');
+      const previewImage = document.getElementById('weird_test') as HTMLImageElement;
 
-    return file;
+      previewImage.src = imgUrl;
+
+      const response = await fetch(imgUrl);
+      if (response) {
+        const blob = await response.blob();
+        const file = new File([blob], 'captureApplication.png', { type: 'image/png' });
+        return file;
+      }
+    }
   } catch (e) {
     console.error(e);
     throw e;
   }
 };
+
+//이미지 다운로드 받는 함수
+function downloadURL(uri: string, name: string) {
+  const link = document.createElement('a');
+  link.download = name;
+  link.href = uri;
+  document.body.appendChild(link);
+  link.click();
+}
